Added hint-claiming and unclaiming
Added a hint-claiming and unclaiming button inside of each row.
The response box is hidden until the user claims the hint.

Made the claimer column in the hint table reference the teamId column
in the team table. This means that all hinters should log in to a unique
admin account.

Wrapped the layout in a SessionProvider so that the `useSession` hook
could be used in client components.

Refactored HintTable into multiple components.